new MAC modal criterion and improved global-local mode checks
