+++ 2_stiffened_panels/2_shear/2_timoshenko_panel.py
@@ -129,7 +129,7 @@ stiff_analysis.post_analysis()
 global_lambda_star = stiff_analysis.min_global_mode_eigenvalue
 
 # predict the actual eigenvalue
-pred_lambda,mode_type = stiff_analysis.predict_crit_load(exy=stiff_analysis.affine_exy)
+pred_lambda,mode_type = stiff_analysis.predict_crit_load(axial=False)
 
 if comm.rank == 0:
     stiff_analysis.print_mode_classification()
