+++ 2_stiffened_panels/2_shear/archive/nasa-verify-ex2.py
