+++ 2_stiffened_panels/2_shear/archive/nasa-verify-ex4.py
