+++ 2_stiffened_panels/archive/1.5_corr_gamma.py
